refactor(runner): move process discovery to lib/process.ts
